+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 ### New features
 * Add parsing of Shape Adjust Values List
+* Add parsing Relative Sizes of shape
 
 ### Fixes
 * Fix misplaced `dxa` and `emu` units of measurements and also fix calculation `dxa` unit